Added new structure of Holoraphic Displays
For the import of Holographic Displays the code has been adapted to recognize the new structure of the "database.yml". Without this change the user gets an internal error so that the code line 43 ("path.location") cannot find a location. In the new structure this has been broken down into individual values.
To be able to import both the old and the new structure, a query was added which can automatically handle both versions